+++ Admin/BaseFieldDescription.php
@@ -11,8 +11,8 @@
 
 namespace Sonata\AdminBundle\Admin;
 
+use Doctrine\Common\Inflector\Inflector;
 use Sonata\AdminBundle\Exception\NoValueException;
-use Symfony\Component\DependencyInjection\Container;
 
 /**
  * A FieldDescription hold the information about a field. A typical
@@ -325,7 +325,7 @@ abstract class BaseFieldDescription implements FieldDescriptionInterface
             return;
         }
 
-        $camelizedFieldName = self::camelize($fieldName);
+        $camelizedFieldName = Inflector::camelize($fieldName);
 
         $getters = array();
         $parameters = array();
@@ -422,10 +422,14 @@ abstract class BaseFieldDescription implements FieldDescriptionInterface
      * @param string $property
      *
      * @return string
+     *
+     * @deprecated Deprecated since version 3.x. Use \Doctrine\Common\Inflector::camelize() instead.
      */
     public static function camelize($property)
     {
-        return Container::camelize($property);
+        @trigger_error('camelize() is deprecated since version 3.x and will be removed in 4.0. Use \Doctrine\Common\Inflector::camelize() instead.', E_USER_DEPRECATED);
+
+        return Inflector::camelize($property);
     }
 
     /**